+++ bot/database/database.py
@@ -2,13 +2,12 @@ import datetime
 
 import motor.motor_asyncio
 
-from config import Config
 
 class Database:
     
-    def __init__(self, uri):
+    def __init__(self, uri, database_name):
         self._client = motor.motor_asyncio.AsyncIOMotorClient(uri)
-        self.db = self._client[Config.SESSION_NAME]
+        self.db = self._client[database_name]
         self.col = self.db.users
