+++ bot/plugins/admin.py
@@ -3,7 +3,7 @@ import os
 
 from pyrogram import Filters
 
-from config import Config
+from ..config import Config
 from ..screenshotbot import ScreenShotBot
