+++ bot/plugins/media.py
@@ -5,7 +5,7 @@ import datetime
 from pyrogram import Filters, InlineKeyboardMarkup, InlineKeyboardButton
 
 from ..utils import is_valid_file, generate_stream_link, get_duration, gen_ik_buttons
-from config import Config
+from ..config import Config
 from ..screenshotbot import ScreenShotBot
