+++ bot/plugins/trim_1.py
@@ -4,7 +4,7 @@ from pyrogram import Filters, ForceReply
 
 from ..utils import sample_fn
 from ..screenshotbot import ScreenShotBot
-from config import Config
+from ..config import Config
 
 
 @ScreenShotBot.on_callback_query(Filters.create(lambda _, query: query.data.startswith('trim')))
