+++ bot/plugins/urls.py
@@ -6,7 +6,7 @@ from pyrogram import Filters, InlineKeyboardMarkup, InlineKeyboardButton
 
 from ..utils import is_url, get_duration, gen_ik_buttons
 from ..screenshotbot import ScreenShotBot
-from config import Config
+from ..config import Config
 
 
 @ScreenShotBot.on_message(Filters.private & Filters.text & Filters.incoming & ~Filters.edited)
