--- run.py
@@ -1,8 +1,0 @@
-import asyncio
-
-from bot import run_bot
-
-
-if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(run_bot())
